sowm: Remove client list while retaining support for 2 workspaces.
Don't track windows in a list; let X do this instead.

`win_prev` and `win_next` are replaced with `win_lower`, which lowers the
current window.

Multiple workspaces are still supported, but there are effectively only two
workspaces. `ws_go` is replaced with `ws_toggle` which toggles the map state
(visibility) of all top-level windows. `win_to_ws` is replaced with `win_hide`,
which unmaps (hides) the current window, effectively moving it to the other
workspace.

This incorporates the "add existing clients" patch as well.